+++ packages/jest-config/src/index.ts
@@ -215,7 +215,6 @@ const groupOptions = (
     testPathIgnorePatterns: options.testPathIgnorePatterns,
     testRegex: options.testRegex,
     testRunner: options.testRunner,
-    testURL: options.testURL,
     timers: options.timers,
     transform: options.transform,
     transformIgnorePatterns: options.transformIgnorePatterns,